[DebugInfo] Check for templated static data member when adding constant to record static fields
Debug info for inline static data members was missing the constant
value,
because the initializer for these static data members is not always
emitted.

Now also try to get the initializer from the templated static data
member if
the variable doesn't have an initializer.

This is an import of https://reviews.llvm.org/D89286